+++ packages/tailwindcss/src/ast.ts
@@ -275,13 +275,18 @@ export function optimizeAst(
 
       // Track used animation names
       if (node.property === 'animation') {
-        for (let keyframeName of extractKeyframeNames(node.value))
+        for (let keyframeName of extractKeyframeNames(node.value)) {
           usedKeyframeNames.add(keyframeName)
+        }
       }
 
       // Create fallback values for usages of the `color-mix(…)` function that reference variables
-      // found in the theme config.
-      if (polyfills & Polyfills.ColorMix && node.value.includes('color-mix(')) {
+      // found in the theme config.
+      if (
+        polyfills & Polyfills.ColorMix &&
+        node.value.includes('color-mix(') &&
+        !context.keyframes
+      ) {
         colorMixDeclarations.get(parent).add(node)
       }